test-requirements.txt
ansible-core
darglint
flake8-docstrings
flake8-quotes
libtmux
lxml
pre-commit
pytest-cov
pytest-mock
pytest-xdist
